refactor: Rename package
Rename the package of the example to avoid Spring conflicts.
