+++ testing/camunda-process-test-example/src/main/java/org/example/workers/ArchiveInvoiceWorker.java
@@ -13,14 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.camunda.workers;
+package org.example.workers;
 
 import com.fasterxml.jackson.databind.JsonNode;
-import io.camunda.services.ArchiveService;
-import io.camunda.services.WiredLegacyException;
 import io.camunda.spring.client.annotation.JobWorker;
 import io.camunda.spring.client.annotation.Variable;
 import io.camunda.spring.client.exception.BpmnError;
+import org.example.services.ArchiveService;
+import org.example.services.WiredLegacyException;
 import org.springframework.stereotype.Component;
 
 @Component
@@ -38,7 +38,7 @@ public class ArchiveInvoiceWorker {
       @Variable("invoice") final JsonNode invoiceJson) {
     try {
       service.archiveInvoice(invoiceId, invoiceJson);
-    } catch (WiredLegacyException e) {
+    } catch (final WiredLegacyException e) {
       throw new BpmnError(
           "LEGACY_ERROR_ARCHIVE", "The archive system had a problem: " + e.getMessage());
     }
